pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paperclaw"
version = "0.1.0"
description = "Research data to academic paper PDF generator with interactive editing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "flask>=3.0",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "weasyprint>=60.0",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "seaborn>=0.12",
]

[project.optional-dependencies]
ai = ["openai>=1.0.0"]
docs = ["python-docx>=1.0.0", "PyPDF2>=3.0.0"]
dev = ["pytest>=7.0", "pytest-cov"]
all = ["openai>=1.0.0", "python-docx>=1.0.0", "PyPDF2>=3.0.0", "pytest>=7.0", "pytest-cov"]

[project.scripts]
paperclaw = "paperclaw.cli:cli"

[tool.setuptools.packages.find]
include = ["paperclaw*"]

[tool.setuptools.package-data]
paperclaw = ["templates/*", "assets/*", "web/templates/*", "web/static/**/*"]
